refactored readColor to use uint8_t

+++ include/pico_apds9960/pico_apds9960.h
@@ -46,8 +46,8 @@ public:
   int readGesture();
 
   int colorAvailable();
-  bool readColor(int& r, int& g, int& b);
-  bool readColor(int& r, int& g, int& b, int& c);
+  bool readColor(uint8_t& r, uint8_t& g, uint8_t& b);
+  bool readColor(uint8_t& r, uint8_t& g, uint8_t& b, uint8_t& c);
 
   int proximityAvailable();
   int readProximity();
